chore: moved sliding function to specific component
